Modernize Python package implementation and notebook.
The Python package was written quickly several years ago using an old package
template.  Many of its features were deprecated at that time.  The package is
now more modern.  Updated the notebook to deal with a soon-to-be deprecated use
of the datetime package.  This is all in preparation for ATPESC 2024.

Followed new instructions for building the package, installing the package, and
running the notebooks to test that procedure.  No issues.  Test of installation
passed as expected.  Reran the notebook for including in the commit and saw no
substantial changes in the nbdiff.